modified:   sonar-project.properties 	modified:   src/v2i-hub/PriorityPlugin/CMakeLists.txt 	modified:   src/v2i-hub/PriorityPlugin/src/PriorityPlugin.cpp 	modified:   src/v2i-hub/PriorityPlugin/src/PriorityPlugin.hpp 	new file:   src/v2i-hub/PriorityPlugin/src/PriorityPluginMessaging.cpp 	modified:   src/v2i-hub/PriorityPlugin/src/PriorityRequestProcessor.cpp 	new file:   src/v2i-hub/PriorityPlugin/src/PriorityRequestProcessor.hpp 	modified:   src/v2i-hub/PriorityPlugin/src/PriorityTypes.hpp 	modified:   src/v2i-hub/PriorityPlugin/test/TestPriority.cpp
