+++ sonar-project.properties
@@ -61,7 +61,8 @@ sonar.modules=    TmxCore, \
   TelematicBridgePlugin, \
   MUSTSensorDriverPlugin, \
   FLIRCameraDriverPlugin, \
-  JSONMessageLoggerPlugin
+  JSONMessageLoggerPlugin, \
+  PriorityPlugin
  
 
 TmxCore.sonar.projectBaseDir                    =src/tmx/TmxCore
@@ -93,6 +94,7 @@ TelematicBridgePlugin.sonar.projectBaseDir     =src/v2i-hub/TelematicBridgePlugi
 MUSTSensorDriverPlugin.sonar.projectBaseDir     =src/v2i-hub/MUSTSensorDriverPlugin
 FLIRCameraDriverPlugin.sonar.projectBaseDir     =src/v2i-hub/FLIRCameraDriverPlugin
 JSONMessageLoggerPlugin.sonar.projectBaseDir =src/v2i-hub/JSONMessageLoggerPlugin
+PriorityPlugin.sonar.projectBaseDir =src/v2i-hub/PriorityPlugin
 
 # C++ Package differences
 # Sources
@@ -135,6 +137,8 @@ MUSTSensorDriverPlugin.sonar.sources  =src
 FLIRCameraDriverPlugin.sonar.sources  =src
 FLIRCameraDriverPlugin.sonar.exclusions =test/**
 JSONMessageLoggerPlugin.sonar.sources =src
+PriorityPlugin.sonar.sources =src
+PriorityPlugin.sonar.exclusions =test/**
 
 # Tests
 # Note: For C++ setting this field does not cause test analysis to occur. It only allows the test source code to be evaluated.
@@ -167,3 +171,4 @@ RSUHealthMonitorPlugin.sonar.tests=test
 TelematicBridgePlugin.sonar.tests=test
 MUSTSensorDriverPlugin.sonar.tests=test
 FLIRCameraDriverPlugin.sonar.tests=test
+PriorityPlugin.sonar.tests=test
